split progress views to separate file
